% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2a.R
\name{convert_to_anndata}
\alias{convert_to_anndata}
\title{s2a}
\usage{
convert_to_anndata(object, assay = "RNA", slot = "data")
}
\arguments{
\item{object}{Seurat object to convert}

\item{assay}{Seurat object assay to add to new object. Default: "RNA"}

\item{slot}{Seurat assay slot to convert. If `data` or `scale.data` is used, 
the `counts` slot (if available) is added to `adata.raw`}
}
\description{
Convert Seurat object to AnnData object
}
